+++ src/main/java/com/starion/createjeicompat/CreateJeiCompatMod.java
@@ -1,22 +1,38 @@
 package com.starion.createjeicompat;
 
 /**
- * Create JEI Compat - A compatibility mod that adds pagination support
- * for Create mod's sequenced assembly recipes in JEI.
- * 
- * This mod extends Create mod (https://github.com/Creators-of-Create/Create)
- * and JEI (https://github.com/mezz/JustEnoughItems) functionality by enhancing
- * the JEI display of sequenced assembly recipes with pagination, allowing
- * recipes with more than 6 steps to be properly displayed.
+ * Pagination for Create sequenced-assembly recipes in JEI and EMI (via JEmi).
+ *
+ * Create: https://github.com/Creators-of-Create/Create
+ * JEI: https://github.com/mezz/JustEnoughItems
+ * EMI: https://github.com/emilyploszaj/emi
  */
+import com.starion.createjeicompat.config.CJCConfigs;
+import com.starion.createjeicompat.emi.EmiScrollHandler;
+import com.starion.createjeicompat.jei.JeiScrollHandler;
+import net.minecraftforge.common.MinecraftForge;
+import net.minecraftforge.eventbus.api.IEventBus;
+import net.minecraftforge.fml.ModList;
+import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod;
+import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 
 @Mod(CreateJeiCompatMod.MOD_ID)
 public class CreateJeiCompatMod {
     public static final String MOD_ID = "createjeicompat";
 
     public CreateJeiCompatMod() {
-        // Register anything here
+        ModLoadingContext modLoadingContext = ModLoadingContext.get();
+        CJCConfigs.register(modLoadingContext);
+
+        IEventBus modBus = FMLJavaModLoadingContext.get().getModEventBus();
+        modBus.register(CJCConfigs.class);
+
+        if (ModList.get().isLoaded("jei")) {
+            MinecraftForge.EVENT_BUS.register(new JeiScrollHandler());
+        }
+        if (ModList.get().isLoaded("emi")) {
+            MinecraftForge.EVENT_BUS.register(new EmiScrollHandler());
+        }
     }
 }
-